requirements.txt
asgiref==3.9.1
Brotli==1.1.0
cffi==2.0.0
cssselect2==0.8.0
Django==5.2.6
django-environ==0.12.0
django-jazzmin==3.0.1
django-widget-tweaks==1.5.0
fonttools==4.60.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.23
pydyf==0.11.0
pyphen==0.17.2
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
